Comment out supervisor v4 config; fix merge error
Some of the configuration in our supervisor config files is
only valid in Supervisor v4 (not yet released); we're still
on v3. Comment out the new directives but include a comment
explaining what's going on.

Remove an old bit of config that had snuck back into
one of the supervisor config files.