POC: install Margarita using pip
This is a proof-of-concept, to stimulate discussion.

With these changes, projects will be able to just
"pip install" a version of Margarita, and easily
update their fab files to rsync that to the remote
systems.

This could make installing pinned versions of
Margarita more convenient since we all know how to
install versions of Python packages.

The only flaw is that it will need to be installed
into the virtualenv that Fabric is using, which
more and more these days is different from the project's
main virtualenv.