Adding `letsencrypt_domains`
Often we'd like to have a letsencrypt cert for multiple domains, e.g. `www.example.com` as well as `example.com`. As currently set up, letsencrypt will certify only the canonical domain at `pillar['domain']` and nothing else.

It's pretty easy to get letsencrypt to do multiple domains: you just include multiple `-d` arguments.

A simple way to achieve this in Margarita is to create a variable `letsencrypt_domains` which can be optionally specified in Pillar as a list of alternative domains (defaulting to the single-item list with the canonical domain).